Hide popup loading indicator if an error/flash is thrown from an AJAX handler (#4364)
Fixes rainlab/builder-plugin#283. Credit to @bennothommo 

When using the popup widget with an AJAX handler, a loading indicator appears whilst the AJAX handler is being run. If an exception or error occurs at this point in time, the exception is shown, but neither the loading indicator nor the backdrop are removed as the popup is technically not 'open' at this stage. This fix should resolve this by tracking the loading state of the popup and removing the backdrop and loading indicator if needed.