Default to using Bootstrap 3 pagination.
L5.6 introduced Bootstrap 4 validation. While it is likely our pagination is all custom and is using the Storm UI, just to be safe, we'll make sure pagination is using Bootstrap 3 classes in all circumstances to retain current functionality.

This should have no effect on simple pagination, which is overriden with our custom template on the next line.