Properly scope the search query in filter widgets.
The search input in filters used a local data request to filter options. While this worked fine for most cases, it failed inside modal relations as the relation keys were not being pushed through due to the request not being linked to the relation modal.

This refactors the request to run through the relation modal, emulating the functionality of "data-track-input" to debounce text entry.

Fixes https://github.com/octobercms/october/issues/4775